broke out client tests into their respective namespaces
