+++ Manatee.Json.Tests/Manatee.Json.Tests.csproj
@@ -81,7 +81,7 @@
     <Compile Include="Patch\JsonPointerResolutionTests.cs" />
     <Compile Include="Patch\TestSuite\JsonPatchTestSuite.cs" />
     <Compile Include="PathExtensions.cs" />
-    <Compile Include="ClientTest.cs" />
+    <Compile Include="ClientTests.cs" />
     <Compile Include="DevTest.cs" />
     <Compile Include="JsonArrayExtensionsTest.cs" />
     <Compile Include="JsonObjectExtensionsTest.cs" />
@@ -89,6 +89,7 @@
     <Compile Include="JsonValueEqualsTest.cs" />
     <Compile Include="JsonValueOperatorTest.cs" />
     <Compile Include="JsonValueParseTest.cs" />
+    <Compile Include="Path\ClientTests.cs" />
     <Compile Include="Path\EvaluationTest.cs" />
     <Compile Include="Path\FilterExpressionTest.cs" />
     <Compile Include="Path\IndexExpressionParseTest.cs" />
@@ -111,6 +112,7 @@
     <Compile Include="Schema\TestSuite\JsonSchemaTestSuite.cs" />
     <Compile Include="Schema\TestSuite\SchemaTest.cs" />
     <Compile Include="Schema\TestSuite\SchemaTestSet.cs" />
+    <Compile Include="Serialization\ClientTests.cs" />
     <Compile Include="Serialization\JsonDeserializerTest.cs" />
     <None Include="app.config" />
     <None Include="Files\baseSchema.json">
